+++ src/main/java/com/example/api_coffeeshop/service/CustomerOrderService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.example.api_coffeeshop.exception.CustomerOrderNotFoundException;
 import com.example.api_coffeeshop.model.CustomerOrder;
 import com.example.api_coffeeshop.repository.CustomerOrderRepository;
 
@@ -17,26 +18,31 @@ public class CustomerOrderService {
     private ItemService itemService;
 
     public CustomerOrder createCustomerOrder(CustomerOrder customerOrder) {
-        CustomerOrder newCustomerOrder = new CustomerOrder();
-        newCustomerOrder.setCustomerName(customerOrder.getCustomerName());
-        return customerOrderRepository.save(newCustomerOrder);
+        return customerOrderRepository.save(customerOrder);
     }
 
+    // TODO: recover items related to that order
+    // TODO: call procedure to compute total
     public CustomerOrder readCustomerOrder(Long id) {
-        return customerOrderRepository.findById(id).orElseThrow();
+        return customerOrderRepository.findById(id)
+                .orElseThrow(() -> new CustomerOrderNotFoundException(id));
     }
 
-    public CustomerOrder updateCustomerOrder(CustomerOrder customerOrder) {
-        CustomerOrder newCustomerOrder = customerOrderRepository.findById(customerOrder.getId()).orElseThrow();
-        newCustomerOrder.setCustomerName(customerOrder.getCustomerName());
-        return customerOrderRepository.save(newCustomerOrder);
+    public CustomerOrder updateCustomerOrder(Long id, CustomerOrder customerOrder) {
+        if (!customerOrderRepository.findById(id).isPresent()) {
+            throw new CustomerOrderNotFoundException(id);
+        }
+        customerOrder.setId(id);
+        return customerOrderRepository.save(customerOrder);
     }
 
-    public CustomerOrder deleteCustomerOrder(Long id) {
-        CustomerOrder newCustomerOrder = customerOrderRepository.findById(id).orElseThrow();
+    public void deleteCustomerOrder(Long id) {
+        if (!customerOrderRepository.findById(id).isPresent()) {
+            throw new CustomerOrderNotFoundException(id);
+        }
+        // Find all items associated with this customer order and remove
         itemService.deleteAllItemByCustomerOrderId(id);
-        customerOrderRepository.delete(newCustomerOrder);
-        return newCustomerOrder;
+        customerOrderRepository.deleteById(id);
     }
 
     public List<CustomerOrder> readAllCustomerOrder() {
